Migrate zkg's configuration to its own class
This moves all directory information into a new Config class that captures zkg's
internal configuration, based on its config file.

The motivation here is that the Manager class does not need to manage this
information, and by making it available from a central configuration, other
classes can also leverage it (as opposed to having to live in the Manager).

It includes various consistency fixes: the zeek_dist path is only used when
actually present, and saving the configuration always uses a canonical
ordering (alphabetically, first by sections, then by options in each section).